Allowing an empty array to call umil->cache_purge.
For the module functions, we will check if !$parent and if not setting $parent to 0 (in case anyone sends '' for the parent it won't give errors).

For the module remove function, I had the automatic remove method backwards so it did not work when you did not send modes.  That is now fixed.

Allowing the mod author to start phpBB before calling UMIL Auto (it checks if defined('IN_PHPBB'), and if it is not it will start the phpBB stuff).

Fixing the current page link so it works for folders outside the root phpbb directory.