I had a good idea for error recording.
Now if there are any errors (when using the frontend/auto versions) a file should be automatically created in umil/error_files and the user will be given notice to download that file and ask the mod author for assistance.

Adding a notice if any permissions are added.

Adding the option for custom functions to return their results (when using the run_actions function).