Fixing the reverse function call for set/unset functions when uninstalling.
Adding acp/common and acp/permissions when not in stand-alone mode.

For the umil_start and result functions, any extra args sent are now checked to see if they are keys in $user->lang and the values of those are used if they exist.

Adding permission_set and permission_unset for adding/removing permissions to groups and roles.